Refactor main JS into modular structure
