+++ site/js/modules/lazyLoader.js
@@ -1,0 +1,31 @@
+/**
+ * Lazy loading
+ * PDFişlemleri.com
+ */
+
+class LazyLoader {
+    constructor() {
+        this.init();
+    }
+
+    init() {
+        const images = document.querySelectorAll('img[data-src]');
+        if (images.length > 0) {
+            const imageObserver = new IntersectionObserver((entries, observer) => {
+                entries.forEach(entry => {
+                    if (entry.isIntersecting) {
+                        const img = entry.target;
+                        img.src = img.dataset.src;
+                        img.classList.remove('lazy');
+                        imageObserver.unobserve(img);
+                    }
+                });
+            });
+
+            images.forEach(img => imageObserver.observe(img));
+        }
+    }
+}
+
+export { LazyLoader };
+
